Refactor favourite button icon selection
This commit refactors the icon selection for the favourite button in the
Favourites control. The icon is now selected based on whether the current page
is a favourite or not. If it is a favourite, the filled heart icon is used,
otherwise the outline heart icon is used. This improves the user experience by
providing a clear visual indication of whether the current page is a favourite or
not.